+++ CHANGES.rst
@@ -9,6 +9,7 @@ Development
 - [ui] Attempt to mitigate https://github.com/marcj/css-element-queries/issues/268
   by upgrading to most recent release
 - [ui] Offer multiple languages when displaying fulltext content
+- [ui] Honor vendor-based data source preference
 
 
 2019-05-15 0.167.2